Update to dynamic grouping of items for analysis.
As discussed with @wendellpiez, aggregating analysis, primarily counts
and simply statistics over more than the context of a given set of the
elements you traditionally use for an assert or report. So how do we do
statistics over the collections at the XML document level, without the
required context as Schematron operates on? In the interim we will make
use of XSL functions to structure XML elements with an `analysis/reports`
path, and soon a function that will structure that data in a report
message (`lv:collect` is the former, the latter TBD).